+++ lib/assembly/grid_function.hpp
@@ -50,7 +50,7 @@ namespace Bempp
 class AssemblyOptions;
 class GeometryFactory;
 class Grid;
-template <int codim> class Entity;
+template <size_t codim> class Entity;
 template <typename BasisFunctionType> class Space;
 
 template <typename BasisFunctionType, typename ResultType>
