+++ lib/assembly/interpolated_function.hpp
@@ -55,9 +55,9 @@ public:
     /** \brief Interpolation grid. */
     const Grid& grid() const;
 
-    virtual int worldDimension() const;
-    virtual int codomainDimension() const;
-    virtual void addGeometricalDependencies(int& geomDeps) const;
+    virtual size_t worldDimension() const;
+    virtual size_t codomainDimension() const;
+    virtual void addGeometricalDependencies(size_t& geomDeps) const;
 
     virtual void evaluate(const Fiber::GeometricalData<CoordinateType>& geomData,
                           arma::Mat<ValueType>& result) const;
